Create FilterTodoItems component
