+++ src/FilterTodoItems.tsx
@@ -1,0 +1,16 @@
+export function FilterTodoItems(props) {
+  const { selected, handleChangeSelect } = props;
+  const optionsOnSelect = [
+    { label: '全て', value: 'all'},
+    { label: '完了', value: 'completed'},
+    { label: '未完了', value: 'uncompleted'},
+  ]
+  
+  return (
+    <select value={selected} onChange={handleChangeSelect}>
+      {optionsOnSelect.map(option => (
+        <option value={option.value} key={option.value}>{option.label}</option>
+      ))}
+    </select>
+  )
+}
